Create initial file structure for Compare pages.
